pre 2.0 monster commit : new dsp code ngvi test

+++ filter.h
@@ -23,5 +23,5 @@
 
 float fir(float *buff,const float *coeff,const int len);
 void  iqfir(float *buff,const float *coeff,const int len,double *I,double *Q);
-float rsfir(float *buff,const float *coeff,const int len ,const double offset ,const double delta);
+float rsfir(double *buff,const float *coeff,const int len ,const double offset ,const double delta);
